pyproject.toml
[project]
name = "ftp-extractor-v2"
dynamic = ["version"]
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "backoff>=2.2.1",
    "datadirtest",
    "freezegun>=1.5.1",
    "ftputil==5.1.0",
    "keboola-component>=1.6.10",
    "keboola-http-client>=1.0.1",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "paramiko>=4.0.0",
    "pydantic>=2.11.3",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "ruff>=0.11.5",
]

[tool.uv.sources]
datadirtest = { git = "https://github.com/keboola/datadirtest" }
